+++ example/table.dart
@@ -31,6 +31,5 @@ Future<void> main() async {
   doc.addFont('Normal', 'swiss', 'Arial', rtf.FontStyle.regular, 9);
   doc.addFont('heading 1', 'swiss', 'Arial', rtf.FontStyle.bold, 14);
   doc.addFont('heading 2', 'swiss', 'Arial', rtf.FontStyle.bold, 12);
-  doc.setHf(rtf.HF.hdLeft, rtf.PageNo());
   await doc.save(File('result.rtf'));
 }