+++ test/rtf_test.dart
@@ -12,13 +12,12 @@ void main() {
       rtf.NewLine(),
       rtf.Text('Second test text', style: rtf.TextStyle(style: 'Normal')),
     ];
-    rtf.Document doc = rtf.Document(el);
+    rtf.Document doc = rtf.Document(el, hdLeft: rtf.PageNo());
     doc.addFont('Normal', 'roman', 'Times New Roman', rtf.FontStyle.regular, 9);
     doc.addFont(
         'heading 1', 'roman', 'Times New Roman', rtf.FontStyle.bold, 14);
     doc.addFont(
         'heading 2', 'roman', 'Times New Roman', rtf.FontStyle.bold, 12);
-    doc.setHf(rtf.HF.hdLeft, rtf.PageNo());
     await doc.save(File('../table.rtf'));
   });
   test('draw a Line', () async {
@@ -36,5 +35,23 @@ void main() {
     doc.addFont(
         'heading 2', 'roman', 'Times New Roman', rtf.FontStyle.bold, 12);
     await doc.save(File('../line.rtf'));
+  });
+  test('draw a Row', () async {
+    rtf.Document doc = rtf.Document([
+      rtf.Section(2, children: List.generate(100, (index) => rtf.Text('kkkk'))),
+      rtf.Column(children: List.generate(100, (index) => rtf.Text('aaaa'))),
+      rtf.Section(1, children: [
+        rtf.Text('start'),
+        rtf.Section(2,
+            children: List.generate(100, (index) => rtf.Text('jjjj'))),
+        rtf.Text('fff')
+      ]),
+    ]);
+    doc.addFont('Normal', 'roman', 'Times New Roman', rtf.FontStyle.regular, 9);
+    doc.addFont(
+        'heading 1', 'roman', 'Times New Roman', rtf.FontStyle.bold, 14);
+    doc.addFont(
+        'heading 2', 'roman', 'Times New Roman', rtf.FontStyle.bold, 12);
+    await doc.save(File('../row.rtf'));
   });
 }